+++ google-cloud-bigtable/src/main/java/com/google/cloud/bigtable/gaxx/grpc/BigtableTransportChannelProvider.java
@@ -23,12 +23,14 @@ import com.google.api.gax.grpc.InstantiatingGrpcChannelProvider;
 import com.google.api.gax.rpc.TransportChannel;
 import com.google.api.gax.rpc.TransportChannelProvider;
 import com.google.auth.Credentials;
+import com.google.cloud.bigtable.data.v2.stub.metrics.ChannelPoolMetricsTracer;
 import com.google.common.base.Preconditions;
 import io.grpc.ManagedChannel;
 import java.io.IOException;
 import java.util.Map;
 import java.util.concurrent.Executor;
 import java.util.concurrent.ScheduledExecutorService;
+import javax.annotation.Nullable;
 
 /**
  * An instance of TransportChannelProvider that provides a TransportChannel through a supplied
@@ -38,10 +40,16 @@ import java.util.concurrent.ScheduledExecutorService;
 public final class BigtableTransportChannelProvider implements TransportChannelProvider {
 
   private final InstantiatingGrpcChannelProvider delegate;
+  private final ChannelPrimer channelPrimer;
+  @Nullable private final ChannelPoolMetricsTracer channelPoolMetricsTracer;
 
   private BigtableTransportChannelProvider(
-      InstantiatingGrpcChannelProvider instantiatingGrpcChannelProvider) {
+      InstantiatingGrpcChannelProvider instantiatingGrpcChannelProvider,
+      ChannelPrimer channelPrimer,
+      ChannelPoolMetricsTracer channelPoolMetricsTracer) {
     delegate = Preconditions.checkNotNull(instantiatingGrpcChannelProvider);
+    this.channelPrimer = channelPrimer;
+    this.channelPoolMetricsTracer = channelPoolMetricsTracer;
   }
 
   @Override
@@ -63,7 +71,8 @@ public final class BigtableTransportChannelProvider implements TransportChannelP
   public BigtableTransportChannelProvider withExecutor(Executor executor) {
     InstantiatingGrpcChannelProvider newChannelProvider =
         (InstantiatingGrpcChannelProvider) delegate.withExecutor(executor);
-    return new BigtableTransportChannelProvider(newChannelProvider);
+    return new BigtableTransportChannelProvider(
+        newChannelProvider, channelPrimer, channelPoolMetricsTracer);
   }
 
   @Override
@@ -75,7 +84,8 @@ public final class BigtableTransportChannelProvider implements TransportChannelP
   public BigtableTransportChannelProvider withHeaders(Map<String, String> headers) {
     InstantiatingGrpcChannelProvider newChannelProvider =
         (InstantiatingGrpcChannelProvider) delegate.withHeaders(headers);
-    return new BigtableTransportChannelProvider(newChannelProvider);
+    return new BigtableTransportChannelProvider(
+        newChannelProvider, channelPrimer, channelPoolMetricsTracer);
   }
 
   @Override
@@ -87,7 +97,8 @@ public final class BigtableTransportChannelProvider implements TransportChannelP
   public TransportChannelProvider withEndpoint(String endpoint) {
     InstantiatingGrpcChannelProvider newChannelProvider =
         (InstantiatingGrpcChannelProvider) delegate.withEndpoint(endpoint);
-    return new BigtableTransportChannelProvider(newChannelProvider);
+    return new BigtableTransportChannelProvider(
+        newChannelProvider, channelPrimer, channelPoolMetricsTracer);
   }
 
   @Deprecated
@@ -101,7 +112,8 @@ public final class BigtableTransportChannelProvider implements TransportChannelP
   public TransportChannelProvider withPoolSize(int size) {
     InstantiatingGrpcChannelProvider newChannelProvider =
         (InstantiatingGrpcChannelProvider) delegate.withPoolSize(size);
-    return new BigtableTransportChannelProvider(newChannelProvider);
+    return new BigtableTransportChannelProvider(
+        newChannelProvider, channelPrimer, channelPoolMetricsTracer);
   }
 
   /** Expected to only be called once when BigtableClientContext is created */
@@ -130,7 +142,14 @@ public final class BigtableTransportChannelProvider implements TransportChannelP
     BigtableChannelPoolSettings btPoolSettings =
         BigtableChannelPoolSettings.copyFrom(delegate.getChannelPoolSettings());
 
-    BigtableChannelPool btChannelPool = BigtableChannelPool.create(btPoolSettings, channelFactory);
+    BigtableChannelPool btChannelPool =
+        BigtableChannelPool.create(btPoolSettings, channelFactory, channelPrimer);
+
+    if (channelPoolMetricsTracer != null) {
+      channelPoolMetricsTracer.registerChannelInsightsProvider(btChannelPool::getChannelInfos);
+      channelPoolMetricsTracer.registerLoadBalancingStrategy(
+          btPoolSettings.getLoadBalancingStrategy().name());
+    }
 
     return GrpcTransportChannel.create(btChannelPool);
   }
@@ -149,12 +168,16 @@ public final class BigtableTransportChannelProvider implements TransportChannelP
   public TransportChannelProvider withCredentials(Credentials credentials) {
     InstantiatingGrpcChannelProvider newChannelProvider =
         (InstantiatingGrpcChannelProvider) delegate.withCredentials(credentials);
-    return new BigtableTransportChannelProvider(newChannelProvider);
+    return new BigtableTransportChannelProvider(
+        newChannelProvider, channelPrimer, channelPoolMetricsTracer);
   }
 
   /** Creates a BigtableTransportChannelProvider. */
   public static BigtableTransportChannelProvider create(
-      InstantiatingGrpcChannelProvider instantiatingGrpcChannelProvider) {
-    return new BigtableTransportChannelProvider(instantiatingGrpcChannelProvider);
+      InstantiatingGrpcChannelProvider instantiatingGrpcChannelProvider,
+      ChannelPrimer channelPrimer,
+      ChannelPoolMetricsTracer outstandingRpcsMetricTracke) {
+    return new BigtableTransportChannelProvider(
+        instantiatingGrpcChannelProvider, channelPrimer, outstandingRpcsMetricTracke);
   }
 }